+++ types/rox-browser/rox-browser-tests.ts
@@ -5,21 +5,21 @@ const flags = {
   superFlag2: new Rox.Flag(),
 };
 
-const variants = {
-  superVariant: new Rox.Variant('value1', ['value1', 'value2']),
+const stringFlags = {
+  strFlag: new Rox.RoxString('value1', ['value1', 'value2']),
 };
 
-const configurations = {
-  superConfiguration: new Rox.Configuration('☀️'),
-  superConfiguration2: new Rox.Configuration(true),
+const numberFlags = {
+  numFlag: new Rox.RoxNumber(12, [18, 24]),
 };
 
 // The register function should be called before the call to Rox.setup()
-Rox.register('default', { ...configurations, ...variants, ...flags });
+Rox.register('default', { ...numberFlags, ...stringFlags, ...flags });
 Rox.setContext({ user: 'John Doe' });
 Rox.setup('ROLLOUT_IO_KEY', {
   impressionHandler,
   configurationFetchedHandler,
+  dynamicPropertyRuleHandler
 }).then(linkTargetGroupAttributes);
 
 Rox.fetch();
@@ -28,6 +28,7 @@ Rox.showOverrides(Rox.RoxOverridesPosition.BottomLeft);
 
 Rox.dynamicApi.isEnabled('system.repotAnalytics', false);
 Rox.dynamicApi.value('ui.textColor', 'red');
+Rox.dynamicApi.getNumber('ui.textSize', 12);
 
 Rox.flags[0].defaultValue;
 Rox.flags[0].name;
@@ -35,15 +36,15 @@ Rox.flags[0].name;
 flags.superFlag.isEnabled();
 flags.superFlag.isEnabled({ user: 'John' });
 
-configurations.superConfiguration.defaultValue;
-configurations.superConfiguration.name;
-configurations.superConfiguration.getValue();
-configurations.superConfiguration.getValue({ user: 'John' });
+numberFlags.numFlag.defaultValue;
+numberFlags.numFlag.name;
+numberFlags.numFlag.getValue();
+numberFlags.numFlag.getValue({ user: 'John' });
 
-variants.superVariant.defaultValue;
-variants.superVariant.name;
-variants.superVariant.getValue();
-variants.superVariant.getValue({ user: 'John' });
+stringFlags.strFlag.defaultValue;
+stringFlags.strFlag.name;
+stringFlags.strFlag.getValue();
+stringFlags.strFlag.getValue({ user: 'John' });
 
 Rox.unfreeze();
 flags.superFlag.unfreeze();
@@ -60,18 +61,16 @@ function linkTargetGroupAttributes() {
   Rox.setCustomNumberProperty('aNumberProperty', 17);
   Rox.setCustomNumberProperty('aNumberProperty', () => 17);
   Rox.setCustomNumberProperty('aNumberProperty', (context: any): number => context.value);
+}
 
-  Rox.setDynamicCustomPropertyRule((propName: string, _context: unknown) => {
-    return propName === 'myPropName';
-  });
+function dynamicPropertyRuleHandler(propName: string, _context: unknown) {
+  return propName === 'myPropName';
 }
 
 function impressionHandler(
-  _reporting: Rox.RoxReporting,
-  _experiment?: Rox.RoxExperiment,
+  _reporting: Rox.RoxReporting
 ) {
-  // If there is no experiment it means that the user has not been enrolled
-  // or that the reporting is not used yet
+  // If _reporting.targeting is false, it mean there were no dashboard conditions, and default value was used
 }
 
 function configurationFetchedHandler(fetcherResult: Rox.RoxFetcherResult) {
