+++ types/rox-browser/v4/index.d.ts
@@ -1,29 +1,26 @@
-// Type definitions for rox-browser 5.0
+// Type definitions for rox-browser 4.8
 // Project: https://rollout.io
 // Definitions by: g-guirado <https://github.com/g-guirado>
-//                 rollout: <https://github.com/rollout>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 3.0
 
 /**
  *
  * Official documentation for rox-browser is available here:
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api
  *
  */
 
 export interface RoxContainer {
-  [key: string]: Flag | RoxNumber | RoxString;
+  [key: string]: Flag | Configuration<any> | Variant;
 }
 
 /**
  * The register function should be called before the call to Rox.setup()
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_register
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_register
  */
 export function register(namespace: string, roxContainer: RoxContainer): void;
-
-export function register(roxContainer: RoxContainer): void;
 
 /**
  * Set Global Context.
@@ -37,17 +34,17 @@ export function setContext(globalContext: unknown): void;
  * Initiate connection with Rox servers for the application identified by the application key.
  * The registered containers will be synced and Rox entities will get the appropriate values.
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_setup
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_setup
  */
 export function setup(apiKey: string, options?: RoxSetupOptions): Promise<unknown>;
 
 export interface RoxSetupOptions {
   version?: string;
-  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_configurationfetchedhandler
+  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_configurationfetchedhandler
   configurationFetchedHandler?(fetcherResult: RoxFetcherResult): void;
   debugLevel?: 'verbose';
-  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_using_the_impressionhandler_option
-  impressionHandler?(reporting: RoxReporting, context: unknown): void;
+  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_using_the_impressionhandler_option
+  impressionHandler?(reporting: RoxReporting, experiment: RoxExperiment, context: unknown): void;
   platform?: string;
   freeze?: RoxFlagFreezeLevel;
   disableNetworkFetch?: boolean;
@@ -58,8 +55,6 @@ export interface RoxSetupOptions {
    * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/debugging/microservices-automated-testing-and-local-development
    */
   roxy?: string;
-  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_dynamicpropertyrulehandler
-  dynamicPropertyRuleHandler?(propName: string, context: any): any;
 }
 
 export enum RoxFetcherStatus {
@@ -67,13 +62,6 @@ export enum RoxFetcherStatus {
   AppliedFromCache = 'APPLIED_FROM_CACHE',
   AppliedFromNetwork = 'APPLIED_FROM_NETWORK',
   ErrorFetchFailed = 'ERROR_FETCH_FAILED'
-}
-
-export enum RoxErrorTrigger {
-  DYNAMIC_PROPERTIES_RULE = 'DYNAMIC_PROPERTIES_RULE',
-  CONFIGURATION_FETCHED_HANDLER = 'CONFIGURATION_FETCHED_HANDLER',
-  IMPRESSION_HANDLER = 'IMPRESSION_HANDLER',
-  CUSTOM_PROPERTY_GENERATOR = 'CUSTOM_PROPERTY_GENERATOR'
 }
 
 export interface RoxFetcherResult {
@@ -86,7 +74,13 @@ export interface RoxFetcherResult {
 export interface RoxReporting {
   name: string;
   value: string;
-  targeting: boolean;
+}
+
+export interface RoxExperiment {
+  identifier: string; //  experiment id
+  name: string;
+  isArchived: boolean;
+  labels: string[]; // experiment's labels. assigned from dashboard
 }
 
 /**
@@ -96,10 +90,8 @@ export interface RoxReporting {
 export function setCustomNumberProperty(name: string, value: number | ((context?: unknown) => number)): void;
 export function setCustomStringProperty(name: string, value: string | ((context?: unknown) => string)): void;
 export function setCustomBooleanProperty(name: string, value: boolean | ((context?: unknown) => boolean)): void;
-
-// https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_setuserspaceunhandlederrorhandler
-export function setUserspaceUnhandledErrorHandler(
-  handler: (errorTrigger: RoxErrorTrigger, error: Error) => void
+export function setDynamicCustomPropertyRule(
+  handler: (propName: string, context: unknown) => number | string | boolean
 ): void;
 
 /**
@@ -107,20 +99,20 @@ export function setUserspaceUnhandledErrorHandler(
  * Calling this function will unfreeze all flags, and using a flag will return it’s most updated value
  *
  * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/feature-flags/flag-freeze#_flag_unfreeze
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_unfreeze_4
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_unfreeze
  */
 export function unfreeze(namespace?: string): void;
 
 /**
  * Pulls the latest configuration and flag values down from the Rollout servers
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_fetch
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_fetch
  */
 export function fetch(): void;
 
 /**
  * Opens the flag override view, providing a debug UI for the application's set of feature flags.
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_showoverrides
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_showoverrides
  */
 export function showOverrides(position?: RoxOverridesPosition): void;
 
@@ -132,7 +124,7 @@ export enum RoxOverridesPosition {
 }
 
 /**
- * Default is none
+ * Default is untilForeground
  *
  * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/feature-flags/flag-freeze
  */
@@ -148,7 +140,7 @@ export interface RoxFlagOptions {
 
 /**
  * Creates a new Flag
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_flag
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_flag
  */
 export class Flag {
   constructor(defaultValue?: boolean, options?: RoxFlagOptions);
@@ -167,46 +159,53 @@ export class Flag {
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined string values
+ * Used to create and manage Rollout feature flags that determine different predefined values
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_roxstring
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_variant
  */
-export class RoxString {
-  constructor(defaultValue: string, options?: ReadonlyArray<string>);
+export class Variant<T extends string = string> {
+  constructor(defaultValue: T, options: ReadonlyArray<T>, name?: string);
 
-  // The name of the RoxString
+  // The name of the Variant
   readonly name: string;
 
-  // Default value of the RoxString
-  readonly defaultValue: string;
+  // Default value of the Variant
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the RoxString, accounting for value overrides
-  getValue(context?: unknown): string;
+  // Returns the current value of the Variant, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 
-  // Unlock the RoxString value from changes from the last time it was freezed
+  // Unlock the Variant value from changes from the last time it was freezed
   unfreeze(): void;
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined number values
+ * manages a remote configuration setting with a value of type string, boolean, or number.
+ * The constructor sets the default value for the remote configuration setting
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_roxnumber
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_configuration
  */
-export class RoxNumber {
-  constructor(defaultValue: number, options?: ReadonlyArray<number>);
+export class Configuration<T extends number | boolean | string> {
+  constructor(defaultValue: T);
 
-  // The name of the RoxNumber
+  // The name of the Configuration
   readonly name: string;
 
-  // Default value of the RoxNumber
-  readonly defaultValue: number;
+  // Default value of the Configuration
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the RoxNumber, accounting for value overrides
-  getValue(context?: unknown): number;
+  // Returns the current value of the Configuration, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 
-  // Unlock the RoxNumber value from changes from the last time it was freezed
+  // Unlock the Configuration value from changes from the last time it was freezed
   unfreeze(): void;
 }
+
+/**
+ * Ensure that TypeScript properly types things with a basic type.
+ * For example, if T is true, returnedtype shall be boolean, not true
+ */
+export type BasicType<T> = T extends boolean ? boolean : T extends number ? number : T extends string ? string : never;
 
 /**
  * Override: Should only be used for development purposes (QA - Feature dev - e2e)
@@ -217,7 +216,7 @@ export class RoxNumber {
  * right after you call Rox.setup. To clear the override from the cache you need to
  * call the Rox.overrides.clearOverride method
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides
  */
 export namespace overrides {
   /**
@@ -226,7 +225,7 @@ export namespace overrides {
    * This function also saves the override value on the local device disk,
    * so it is "remembered" for the next the SDK is loaded to production.
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_setoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_setoverride
    *
    * Note that for boolean flag we still give the value as a string.
    */
@@ -235,7 +234,7 @@ export namespace overrides {
   /**
    * Clears the override value from the flag (and the disk).
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_clearoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_clearoverride
    */
   function clearOverride(nameSpacedFlagName: string): void;
 
@@ -249,7 +248,7 @@ export namespace overrides {
   /**
    * full flag name including namespace
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_hasoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_hasoverride
    */
   function hasOverride(nameSpacedFlagName: string): boolean;
 }
@@ -257,7 +256,7 @@ export namespace overrides {
 /**
  * Dynamic API is an alternative to Rollout static API for defining flags on the
  * different container objects and accessing them from that container object.
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_dynamicapi
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/feature-flags/dynamic-api
  */
 export namespace dynamicApi {
   /**
@@ -266,14 +265,9 @@ export namespace dynamicApi {
   function isEnabled(nameSpacedFlagName: string, defaultValue: boolean, context?: unknown): boolean;
 
   /**
-   * Getting string value of a string flag
+   * Getting string value of a Variant flag
    */
   function value(nameSpacedFlagName: string, defaultValue: string, context?: unknown): string;
-
-  /**
-   * Getting string value of a number flag
-   */
-  function getNumber(nameSpacedFlagName: string, defaultValue: number, context?: unknown): number;
 }
 
 export const flags: ReadonlyArray<Flag>;
