+++ types/rox-node/v4/index.d.ts
@@ -1,35 +1,32 @@
-// Type definitions for rox-node 5.0
+// Type definitions for rox-node 4.8
 // Project: https://rollout.io
 // Definitions by: g-guirado <https://github.com/g-guirado>
-//                 rollout: <https://github.com/rollout>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 3.0
 
 /**
  *
  * Official documentation for rox-node is available here:
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api
  *
  */
 
 export interface RoxContainer {
-  [key: string]: Flag | RoxNumber | RoxString;
+  [key: string]: Flag | Configuration<any> | Variant;
 }
 
 /**
  * The register function should be called before the call to Rox.setup()
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_register
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_register
  */
 export function register(namespace: string, roxContainer: RoxContainer): void;
-
-export function register(roxContainer: RoxContainer): void;
 
 /**
  * Set Global Context.
  * You can think of Global Context as a default context
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_setcontext
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_setcontext
  */
 export function setContext(globalContext: unknown): void;
 
@@ -37,7 +34,7 @@ export function setContext(globalContext: unknown): void;
  * Initiate connection with Rox servers for the application identified by the application key.
  * The registered containers will be synced and Rox entities will get the appropriate values.
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_setup
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_setup
  */
 export function setup(apiKey: string, options?: RoxSetupOptions): Promise<unknown>;
 
@@ -46,20 +43,18 @@ export interface RoxSetupOptions {
   // https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/reporting/configuration-fetched-handler
   configurationFetchedHandler?(fetcherResult: RoxFetcherResult): void;
   debugLevel?: 'verbose';
-  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_using_the_impressionhandler_option
-  impressionHandler?(reporting: RoxReporting, context: unknown): void;
+  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_using_the_impressionhandler_option
+  impressionHandler?(reporting: RoxReporting, experiment: RoxExperiment, context: unknown): void;
   platform?: string;
   fetchIntervalInSec?: number;
   disableNetworkFetch?: boolean;
   devModeSecret?: string;
-  /**
-   * Set Roxy's URL for automated tests or local development.
-   *
+  /**	
+   * Set Roxy's URL for automated tests or local development.	
+   *	
    * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/debugging/microservices-automated-testing-and-local-development
-   */
+   */	
   roxy?: string;
-  // https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_dynamicpropertyrulehandler
-  dynamicPropertyRuleHandler?(propName: string, context: any): any;
 }
 
 export enum RoxFetcherStatus {
@@ -67,13 +62,6 @@ export enum RoxFetcherStatus {
   AppliedFromCache = 'APPLIED_FROM_CACHE',
   AppliedFromNetwork = 'APPLIED_FROM_NETWORK',
   ErrorFetchFailed = 'ERROR_FETCH_FAILED'
-}
-
-export enum RoxErrorTrigger {
-  DYNAMIC_PROPERTIES_RULE = 'DYNAMIC_PROPERTIES_RULE',
-  CONFIGURATION_FETCHED_HANDLER = 'CONFIGURATION_FETCHED_HANDLER',
-  IMPRESSION_HANDLER = 'IMPRESSION_HANDLER',
-  CUSTOM_PROPERTY_GENERATOR = 'CUSTOM_PROPERTY_GENERATOR'
 }
 
 export interface RoxFetcherResult {
@@ -86,7 +74,13 @@ export interface RoxFetcherResult {
 export interface RoxReporting {
   name: string;
   value: string;
-  targeting: boolean;
+}
+
+export interface RoxExperiment {
+  identifier: string; //  experiment id
+  name: string;
+  isArchived: boolean;
+  labels: string[]; // experiment's labels. assigned from dashboard
 }
 
 /**
@@ -95,22 +89,20 @@ export interface RoxReporting {
 export function setCustomNumberProperty(name: string, value: number | ((context?: unknown) => number)): void;
 export function setCustomStringProperty(name: string, value: string | ((context?: unknown) => string)): void;
 export function setCustomBooleanProperty(name: string, value: boolean | ((context?: unknown) => boolean)): void;
-
-// https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_setuserspaceunhandlederrorhandler
-export function setUserspaceUnhandledErrorHandler(
-  handler: (errorTrigger: RoxErrorTrigger, error: Error) => void
+export function setDynamicCustomPropertyRule(
+  handler: (propName: string, context: unknown) => number | string | boolean
 ): void;
 
 /**
  * Pulls the latest configuration and flag values down from the Rollout servers
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_fetch
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_fetch
  */
 export function fetch(): void;
 
 /**
  * Creates a new Flag
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_flag
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_flag
  */
 export class Flag {
   constructor(defaultValue?: boolean);
@@ -126,40 +118,47 @@ export class Flag {
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined string values
+ * Used to create and manage Rollout feature flags that determine different predefined values
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_rox_roxstring
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_variant
  */
-export class RoxString {
-  constructor(defaultValue: string, options?: ReadonlyArray<string>);
+export class Variant<T extends string = string> {
+  constructor(defaultValue: T, options: ReadonlyArray<T>, name?: string);
 
-  // The name of the string flag
+  // The name of the Variant
   readonly name: string;
 
-  // Default value of the string flag
-  readonly defaultValue: string;
+  // Default value of the Variant
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the string flag, accounting for value overrides
-  getValue(context?: unknown): string;
+  // Returns the current value of the Variant, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined number values
+ * manages a remote configuration setting with a value of type string, boolean, or number.
+ * The constructor sets the default value for the remote configuration setting
  *
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_rox_roxnumber
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/nodejs-api#_configuration
  */
-export class RoxNumber {
-  constructor(defaultValue: number, options?: ReadonlyArray<number>);
+export class Configuration<T extends number | boolean | string> {
+  constructor(defaultValue: T);
 
-  // The name of the number flag
+  // The name of the Configuration
   readonly name: string;
 
-  // Default value of the number flag
-  readonly defaultValue: number;
+  // Default value of the Configuration
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the number flag, accounting for value overrides
-  getValue(context?: unknown): number;
+  // Returns the current value of the Configuration, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 }
+
+/**
+ * Ensure that TypeScript properly types things with a basic type.
+ * For example, if T is true, returnedtype shall be boolean, not true
+ */
+export type BasicType<T> = T extends boolean ? boolean : T extends number ? number : T extends string ? string : never;
 
 /**
  * Override: Should only be used for development purposes (QA - Feature dev - e2e)
@@ -171,7 +170,7 @@ export class RoxNumber {
  * call the Rox.overrides.clearOverride method.
  *
  * One can refer to the javascript-browser-api for this feature:
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides
  */
 export namespace overrides {
   /**
@@ -180,7 +179,7 @@ export namespace overrides {
    * This function also saves the override value on the local device disk,
    * so it is "remembered" for the next the SDK is loaded to production.
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_setoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_setoverride
    *
    * Note that for boolean flag we still give the value as a string.
    */
@@ -189,7 +188,7 @@ export namespace overrides {
   /**
    * Clears the override value from the flag (and the disk).
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_clearoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_clearoverride
    */
   function clearOverride(nameSpacedFlagName: string): void;
 
@@ -203,7 +202,7 @@ export namespace overrides {
   /**
    * full flag name including namespace
    *
-   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/javascript-browser-api#_rox_overrides_hasoverride
+   * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/4.9/api-reference/javascript-browser-api#_rox_overrides_hasoverride
    */
   function hasOverride(nameSpacedFlagName: string): boolean;
 }
@@ -211,7 +210,7 @@ export namespace overrides {
 /**
  * Dynamic API is an alternative to Rollout static API for defining flags on the
  * different container objects and accessing them from that container object.
- * https://docs.cloudbees.com/docs/cloudbees-feature-flags-api/latest/api-reference/nodejs-api#_rox_dynamicapi
+ * https://docs.cloudbees.com/docs/cloudbees-feature-flags/latest/feature-flags/dynamic-api
  */
 export namespace dynamicApi {
   /**
@@ -220,14 +219,9 @@ export namespace dynamicApi {
   function isEnabled(nameSpacedFlagName: string, defaultValue: boolean, context?: unknown): boolean;
 
   /**
-   * Getting string value of a string flag
+   * Getting string value of a Variant flag
    */
   function value(nameSpacedFlagName: string, defaultValue: string, context?: unknown): string;
-
-  /**
-   * Getting string value of a number flag
-   */
-  function getNumber(nameSpacedFlagName: string, defaultValue: number, context?: unknown): number;
 }
 
 export const flags: ReadonlyArray<Flag>;
