+++ types/rox-react-native/v4/rox-react-native-tests.ts
@@ -5,20 +5,20 @@ const flags = {
     superFlag2: new Rox.Flag(),
 };
 
-const stringFlags = {
-    strFlag: new Rox.RoxString('value1', ['value1', 'value2']),
+const variants = {
+    superVariant: new Rox.Variant('value1', ['value1', 'value2'])
 };
 
-const numberFlags = {
-    numFlag: new Rox.RoxNumber(12, [18, 24]),
+const configurations = {
+    superConfiguration: new Rox.Configuration('☀️'),
+    superConfiguration2: new Rox.Configuration(true),
 };
 
 // The register function should be called before the call to Rox.setup()
-Rox.register('default', { ...stringFlags, ...numberFlags, ...flags });
+Rox.register('default', { ...configurations, ...variants, ...flags });
 Rox.setup('ROLLOUT_IO_KEY', {
     impressionHandler,
     configurationFetchedHandler,
-    dynamicPropertyRuleHandler
 }).then(linkTargetGroupAttributes);
 
 Rox.dynamicApi.isEnabled('system.repotAnalytics', false);
@@ -29,13 +29,13 @@ Rox.flags[0].name;
 
 flags.superFlag.isEnabled();
 
-numberFlags.numFlag.defaultValue;
-numberFlags.numFlag.name;
-numberFlags.numFlag.getValue();
+configurations.superConfiguration.defaultValue;
+configurations.superConfiguration.name;
+configurations.superConfiguration.getValue();
 
-stringFlags.strFlag.defaultValue;
-stringFlags.strFlag.name;
-stringFlags.strFlag.getValue();
+variants.superVariant.defaultValue;
+variants.superVariant.name;
+variants.superVariant.getValue();
 
 Rox.unfreeze();
 flags.superFlag.unfreeze();
@@ -49,17 +49,16 @@ function linkTargetGroupAttributes() {
 
     Rox.setCustomNumberProperty('aNumberProperty', 17);
     Rox.setCustomNumberProperty('aNumberProperty', () => 17);
+
+    Rox.setDynamicCustomPropertyRule((propName: string, _context: unknown) => {
+        return propName === 'myPropName';
+    });
 }
 
-function dynamicPropertyRuleHandler(propName: string, _context: unknown) {
-    return propName === 'myPropName';
+function impressionHandler(_reporting: Rox.RoxReporting, _experiment?: Rox.RoxExperiment) {
+    // If there is no experiment it means that the user has not been enrolled
+    // or that the reporting is not used yet
 }
-
-function impressionHandler(
-    _reporting: Rox.RoxReporting
-  ) {
-    // If _reporting.targeting is false, it mean there were no dashboard conditions, and default value was used
-  }
 
 function configurationFetchedHandler(fetcherResult: Rox.RoxFetcherResult) {
     if (fetcherResult.hasChanges && fetcherResult.fetcherStatus === Rox.RoxFetcherStatus.AppliedFromCache) {
